requirements.txt
alembic==1.8.0
anyio==3.6.1
asgiref==3.5.2
asn1crypto==1.5.1
bcrypt==3.2.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
cryptography==36.0.2
dnspython==2.2.1
ecdsa==0.17.0
email-validator==1.2.1
fastapi==0.78.0
greenlet==1.1.2
h11==0.13.0
httptools==0.4.0
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.1
MarkupSafe==2.1.1
mysql-connector-python==8.0.29
orjson==3.7.5
oscrypto==1.3.0
passlib==1.7.4
protobuf==4.21.2
pyasn1==0.4.8
pycparser==2.21
pycryptodomex==3.15.0
pydantic==1.9.1
PyJWT==2.4.0
pyOpenSSL==22.0.0
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.1
PyYAML==6.0
requests==2.28.1
rsa==4.8
six==1.16.0
sniffio==1.2.0
snowflake-connector-python==2.7.9
snowflake-sqlalchemy==1.3.4
SQLAlchemy==1.4.39
starlette==0.19.1
typing_extensions==4.2.0
ujson==5.3.0
urllib3==1.26.9
uvicorn==0.17.6
watchgod==0.8.2
websockets==10.3
zipp==3.8.0
